requirements.txt
Flask==3.0.0
requests
pandas
qrcode
Pillow
PyJWT
gunicorn